refactor(car): Rename racing car package
