+++ src/main/java/racingcar/MovingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar;
 
 @FunctionalInterface
 public interface MovingStrategy {
