+++ src/test/java/racingcar/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
